README File Edit
Brittany D. Wilson
bdwilson@my.waktech.edu
My Github account is bigdata10101
The repository I created is named bigdata10101 and is linked to my Assignments folder on my hard drive.
To clone, I opened Github on desktop, I was prompted with cloning a repository so I selected that option and I believe it was created successfully. I'm not quite sure how this works and where I'm supposed to make the edit.